requirements.txt
gmplot==1.2.0
osmium==2.14.3
tqdm==4.25.0
geopy==1.16.0
SRTM.py==0.3.4
pytest==3.7.4
networkx==2.1
Django==2.1.1
djangorestframework==3.8.2
pytest-django==3.4.2
googlemaps==3.0.2
